I started modifying the CSS to make it look like I want it to look.
I have changed the standard colours to be like the smileys.. I feel the focus is more desitrubing than helpfull, and are now going to remove it. This is also done because of the way it is implemented. It's current implementation only works on two buttons, which does not help when the primary purpose is now to have three buttons