Started working on a non prompt version
I started working on a non prompt version of feedback.js
There are still a glitch in the prompt version, since the images loads way to slowly. To circumvent this, I created new smileys in CSS, and which can be target with Jqueury, so it should be able to work at some point..

To work with the files without causing to much frustration I made seperate CSS files, a seperate JavaScript file and html.. This way I can modify as I want, without causing problems with the prompt version.

I'm considering using SVG files instead of both CSS and the current png files, because it makes the code easier to read, and is more effective than png