Started working on the custom feedbacklog
I have now implemented the basic functionaliy of the custom feedback log.. This is going to be used as a non-interrupting feedback gateway, as opposed by the prompt version.

At the current moment, the feedback is shown, the custom listeners is attached, and the images is shown correctly..

The css is not working properly though, which destroys the ilusion of animation. The detailed log is not working either.