Basic feedback log functionality is implemented
Now the log feedback almost works as it is supposed to.. It can handle both detailed feedback and normal feedback. There is a bug with the no thanks button, I currently don't understand, and there are some basic flaws in the CSS.

A a later point I will add functionality to say where the log should be placed (bottom-left, bottom-right, right, left), add inline functionality so the user can place it whereever he/she wants in the page and I will begin work on a prompt version.

The prompt version should be used to ask multiple questions at once. This way it is ensured that the same person answers all questions or none, which can be usefull in some situations..