Fixed the CSS
I have now fixed the CSS, and the look of the log is quite decent.. Many improvements can still be made, but this version should be able to be used in many instances.

the cancel button still does not work, but I haven't looked at it yet.. Next goal will be to include options for positioning of the log

r your changes. Lines starting