Added functionality for placement of the log
I have now added functionality for placing the feedback logs 4 places on the screen. This is done by giving the boxes a unique ID and targeting them with dedicated CSS based on the ID..

Because of this I have restructured the CSS to make it clear which CSS is global, and which CSS only target one placement. This way users can simply delete the CSS if they only want one layout.

I have also updated the example file to show of this new funcitonality

BUGS:
There is currently a bug where the log in the bottom-right corner destroys the cunftionality of the other boxes after which a refresh of the page is necessary