Started listing the different funcitonality that would be cool to have
I have divided the src folder into 4 different files, which should each be of different functionality.

FeedbackContact should enable the user to send questions easily to the company

FedbackFeeling is the current available feature, where the user can indicates his emotions as an aswer to a specific question

FeedbackPrompt should be a prompt that contains a list of questions and should prompt the user for feedback.

FeedbackYesNo should have much the same functionality as FeedbackFeeling, but only contain yes/no button, with the option to display the yes/no as thumbs up / thumbs down.

FeedbackYesNo will be the next functionality I work on.